feat: get_total_paragraph_count 함수 추가 및 구조체 카운팅 변수 타입 변경(int -> size_t)

+++ Assignment3/document_analyzer.h
@@ -27,17 +27,17 @@ size_t get_sentence_word_count(const char** sentence);
 int print_as_tree(const char* filename);
 
 typedef struct {
-    int word_count;
+    size_t word_count;
     char** words;
 } sentence_t;
 
 typedef struct {
-    int sentence_count;
+    size_t sentence_count;
     sentence_t* sentences;
 } paragraph_t;
 
 typedef struct {
-    int paragraph_count;
+    size_t paragraph_count;
     paragraph_t* paragraphs;
 } document_t;
